requirements.txt
python-telegram-bot==20.8
flask==2.3.3
google-generativeai==0.4.1